api/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
google-cloud-pubsub==2.32.0

